fix(sqlite): recover request stats after contention
